% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget.R
\name{cbudget}
\alias{cbudget}
\title{Carbonate Budget}
\usage{
cbudget(pp, sp, pe, se)
}
\arguments{
\item{pp}{object of class \code{carb} pertaining to primary production}

\item{sp}{object of class \code{carb} pertaining to secondary production}

\item{pe}{object of class \code{carb} pertaining to primary erosion}

\item{se}{object of class \code{carb} pertaining to secondary erosion}
}
\value{
carbonate budget (net kg/m2yr)
}
\description{
This function combines outputs of \code{pproduction}, \code{sproduction},
\code{perosion} & \code{serosion} to calculate overall carbonate budget
}
\examples{



}
