% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erosion.R
\name{perosion}
\alias{perosion}
\title{Primary Erosion}
\usage{
perosion(TL, species, ta)
}
\arguments{
\item{TL}{numeric vector of individual lengths}

\item{species}{character vector of species codes as per AIMS convention for each individual}

\item{ta}{transect area (m2)}
}
\value{
total primary erosion (kg/m2yr)
}
\description{
This function uses fish survey data to quantify erosion (kg/m2yr)
}
