% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pproduction.R
\name{pproduction}
\alias{pproduction}
\title{Primary Production}
\usage{
pproduction(species, cover, lat)
}
\arguments{
\item{species}{character vector of coral types as per the below naming convention}

\item{cover}{numeric vector of percent cover (0-100) of each coral type}

\item{lat}{numeric vector of latitude}
}
\value{
total primary production (kg/m2yr)
}
\description{
This function uses coral cover data to calculate carbonate production (kg/m2yr)
}
\details{
This function currently supports parameterisation for 56 coral taxa as below:\tabular{ll}{
   \strong{Taxa} \tab \strong{Naming Convention} \cr
   \emph{Acanthastrea} spp \tab "ACANTHASTREA" \cr
   \emph{Acropora} spp (arborescent) \tab "ACB" \cr
   \emph{Acropora} spp (digitate) \tab "ACD" \cr
   \emph{Acropora} spp (encrusting) \tab "ACE" \cr
   \emph{Acropora} spp (corymbose) \tab "ACO" \cr
   \emph{Acropora} spp (submassive) \tab "ACS" \cr
   \emph{Acropora} spp (table) \tab "ACT" \cr
   \emph{Acropora} spp (bottlebrush) \tab "ACX" \cr
   \emph{Alveopora} spp \tab "ALVEOPORA" \cr
   \emph{Astreopora} spp \tab "ASTREOPORA" \cr
   \emph{Caulastrea} spp \tab "CAULASTREA" \cr
   \emph{Coscinaraea} spp \tab "COSCINARAEA" \cr
   \emph{Cyphastrea} spp \tab "CYPHASTREA" \cr
   \emph{Diploastrea} spp \tab "DIPLOASTREA" \cr
   \emph{Echinophyllia/Oxypora} (indeterminate) \tab "ECHINOPHYLLIA.OXYPORA" \cr
   \emph{Echinophyllia} spp \tab "ECHINOPHYLLIA" \cr
   \emph{Echinopora} spp \tab "ECHINOPORA" \cr
   \emph{Euphyllia} spp \tab "EUPHYLLIA" \cr
   \emph{Favia} spp \tab "FAVIA" \cr
   \emph{Favid} spp \tab "FAVID.SPP" \cr
   \emph{Favites} spp \tab "FAVITES" \cr
   \emph{Fungiidae} (solitary) \tab "SOLITARY.FUNGID" \cr
   \emph{Galaxea} spp \tab "GALAXEA" \cr
   \emph{Gardineroseris} spp \tab "GARDINEROSERIS" \cr
   \emph{Goniastrea} spp \tab "GONIASTREA" \cr
   \emph{Goniopora} spp \tab "GONIOPORA" \cr
   \emph{Hydnophora} spp \tab "HYDNOPHORA" \cr
   \emph{Isopora} spp \tab "ISOPORA" \cr
   \emph{Leptastrea} spp \tab "LEPTASTREA" \cr
   \emph{Leptoria} spp \tab "LEPTORIA" \cr
   \emph{Leptoseris} spp \tab "LEPTOSERIS" \cr
   \emph{Lopbophyllia} spp \tab "LOBOPHYLLIA" \cr
   \emph{Merulina} spp \tab "MERULINA" \cr
   \emph{Montastrea} spp \tab "MONTASTREA" \cr
   \emph{Montipora} spp \tab "MONTIPORA" \cr
   \emph{Mycedium} spp \tab "MYCEDIUM" \cr
   \emph{Oulophyllia} spp \tab "OULOPHYLLIA" \cr
   \emph{Oxypora} spp \tab "OXYPORA" \cr
   \emph{Pachyseris} spp \tab "PACHYSERIS" \cr
   \emph{Palauastrea} spp \tab "PALAUASTREA" \cr
   \emph{Paraclavarina} spp \tab "PARACLAVARINA" \cr
   \emph{Pavona} spp \tab "PAVONA" \cr
   \emph{Pectinia} spp \tab "PECTINIA" \cr
   \emph{Physogyra} spp \tab "PHYSOGYRA" \cr
   \emph{Platygyra} spp \tab "PLATYGYRA" \cr
   \emph{Plerogyra} spp \tab "PLEROGYRA" \cr
   \emph{Plesiastrea} spp \tab "PLESIASTREA" \cr
   \emph{Pocillopora} spp \tab "POCILLOPORA" \cr
   \emph{Podabacia} spp \tab "PODABACIA" \cr
   \emph{Porites} spp \tab "PORITES" \cr
   \emph{Psammocora} spp \tab "PSAMMOCORA" \cr
   \emph{Seriatopora} spp \tab "SERIATOPORA" \cr
   \emph{Stylocoeniella} spp \tab "STYLOCOENIELLA" \cr
   \emph{Stylophora} spp \tab "STYLOPHORA" \cr
   \emph{Symphyllia} spp \tab "SYMPHYLLIA" \cr
   \emph{Turbinaria} spp \tab "TURBINARIA" \cr
}
}
