% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/production.R
\name{sproduction}
\alias{sproduction}
\title{Secondary Production}
\usage{
sproduction(species, cover, region, shelf)
}
\arguments{
\item{cover}{numeric vector of percentage cover of each substrate type}

\item{shelf}{character vector of shelf position (I = inner, M = mid-shelf, O = outer)}

\item{substrate}{character vector of algal substrate types}

\item{rug}{numeric vector of chain rugosity}
}
\value{
total secondary production (kg/m2yr)
}
\description{
This function calculates carbonate production (kg/m2yr) of all secondary
(non-coral) reef components, namely, calcareous algae
}
\examples{



}
