% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perosion.R
\name{perosion}
\alias{perosion}
\title{Primary Erosion}
\usage{
perosion(TL, species, ta)
}
\arguments{
\item{TL}{numeric vector of individual body length (cm)}

\item{species}{character vector of species codes as per the below naming convention}

\item{ta}{transect area (m2)}
}
\value{
total primary erosion (kg/m2yr)
}
\description{
This function uses fish survey data to quantify bioerosion (kg/m2yr)
}
\details{
This function currently supports parameterisation for 23 fish species as below:\tabular{ll}{
   \strong{Species} \tab \strong{Naming Convention} \cr
   \emph{Bolbometopon muricatum} \tab "BOL.MURI" \cr
   \emph{Calotomus carolinus} \tab "CAL.CARO" \cr
   \emph{Cetoscarus bicolor} \tab "CET.BIC" \cr
   \emph{Chlorurus bleekeri} \tab "CHS.BLEE" \cr
   \emph{Chlorurus frontalis} \tab "CHS.FRON" \cr
   \emph{Chlorurus japanensis} \tab "CHS.JAPA" \cr
   \emph{Chlorurus mircrorhinos} \tab "CHS.MICR" \cr
   \emph{Chlorurus sordidus} \tab "CHS.SORD" \cr
   \emph{Hipposcarus longiceps} \tab "HIP.LONG" \cr
   \emph{Scarus altipinnis} \tab "SCA.ALTI" \cr
   \emph{Scarus chameleon} \tab "SCA.CHAM" \cr
   \emph{Scarus dimidatus} \tab "SCA.DIMI" \cr
   \emph{Scarus flavipectoralis} \tab "SCA.FLAV" \cr
   \emph{Scarus forsteni} \tab "SCA.FORS" \cr
   \emph{Scarus frenatus} \tab "SCA.FREN" \cr
   \emph{Scarus ghobban} \tab "SCA.GHOB" \cr
   \emph{Scarus globiceps} \tab "SCA.GLOB" \cr
   \emph{Scarus longipinnis} \tab "SCA.LONG" \cr
   \emph{Scarus niger} \tab "SCA.NIGR" \cr
   \emph{Scaru oviceps} \tab "SCA.OVIC" \cr
   \emph{Scarus psittacus} \tab "SCA.PSIT" \cr
   \emph{Scarus rivulatus} \tab "SCA.RIVU" \cr
   \emph{Scarus rubroviolaceus} \tab "SCA.RUBR" \cr
   \emph{Scarus schlegeli} \tab "SCA.SCHL" \cr
   \emph{Scarus spinus} \tab "SCA.SPIN" \cr
   \emph{Scarus quoyi} \tab "SCA.QUO" \cr
}
}
