% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/production.R
\name{pproduction}
\alias{pproduction}
\title{Primary Production}
\usage{
pproduction(species, cover, lat)
}
\arguments{
\item{species}{character vector of species names as per AIMS convention}

\item{cover}{numeric vector of percent cover of each species}

\item{lat}{numeric vector of latitude}
}
\value{
total primary production (kg/m2yr)
}
\description{
This function uses coral cover data to calculate carbonate production (kg/m2yr)
}
