% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serosion.R
\name{serosion}
\alias{serosion}
\title{Secondary Erosion}
\usage{
serosion(cover, rug, shelf)
}
\arguments{
\item{cover}{numeric vector of percentage cover (0-100) of non-coral substrate}

\item{rug}{habitat rugosity score (0-5)}

\item{shelf}{character vector of shelf position ("I" = inner, "M" = mid-shelf, "O" = outer)}
}
\value{
total secondary erosion (kg/m2yr)
}
\description{
This function calculates total erosion (kg/m2yr) of all secondary (non-fish)
reef components, namely, micro- and macro-borers
}
