% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sproduction.R
\name{sproduction}
\alias{sproduction}
\title{Secondary Production}
\usage{
sproduction(species, cover, rug, chain = TRUE, region, shelf)
}
\arguments{
\item{species}{character vector of algal types as per the below naming convention}

\item{cover}{numeric vector of percentage cover (0-100) of each algal type}

\item{rug}{habitat rugosity (either chain-link measurement or visual 0-5 scale)}

\item{chain}{logical if rugosity is chain-link, default is TRUE}

\item{region}{character vector of GBR region ("north", "central" or "south")}

\item{shelf}{character vector of shelf position ("I" = inner, "M" = mid-shelf, "O" = outer)}
}
\value{
total secondary production (kg/m2yr)
}
\description{
This function calculates carbonate production (kg/m2yr) of all secondary
(non-coral) reef components, namely, calcareous algae
}
\details{
This function currently supports parameterisation of 4 calcifying algal taxa as listed below:\tabular{ll}{
   \strong{Taxa} \tab \strong{Naming Convention} \cr
   Crustose corraline algae (\emph{Corallinaceae}) \tab "CCA" \cr
   \emph{Halimeda} spp \tab "Halimeda' \cr
   Calcareous red macroalgae (\emph{Amphiroa} spp) \tab "Calcareous.Red.Macroalgae" \cr
   \emph{Peyssonnelia} spp \tab "Peyssonnelia" \cr
}
}
