% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/production.R
\name{sproduction}
\alias{sproduction}
\title{Secondary Production}
\usage{
sproduction(species, cover, rug, region, shelf)
}
\arguments{
\item{species}{character vector of algal types}

\item{cover}{numeric vector of percentage cover of each algal type}

\item{rug}{numeric vector of chain rugosity}

\item{region}{character vector of GBR region ("north", "central" or "south")}

\item{shelf}{character vector of shelf position ("I" = inner, "M" = mid-shelf, "O" = outer)}
}
\value{
total secondary production (kg/m2yr)
}
\description{
This function calculates carbonate production (kg/m2yr) of all secondary
(non-coral) reef components, namely, calcareous algae
}
